backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
insightface>=0.7.3
onnxruntime>=1.16.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
aiofiles>=23.2.1
python-dotenv>=1.0.0
msal>=1.24.0
python-jose[cryptography]>=3.3.0
